Moved convertOperationsToSuperRootFieldsInAST
